Factor out base.
Now __init__.py can be cleaned up.  Import some variables for backward
compatibility.  For the one private one only used in a test, change the
test.  Change imports on top of cartesian, etc., to be from .base.
